+++ src/algorithms/tracking/adapters/glonass_l2_ca_dll_pll_tracking.h
@@ -3,6 +3,8 @@
  * \brief  Interface of an adapter of a DLL+PLL tracking loop block
  * for Glonass L2 C/A to a TrackingInterface
  * \author Damian Miralles, 2018, dmiralles2009(at)gmail.com
+ * \author Javier Arribas, 2025 javier.arribas(at)cttc.es
+ * \author Carles Fernandez-Prades, 2025 carles.fernandez(at)cttc.es
  *
  *
  * Code DLL + carrier PLL according to the algorithms described in:
@@ -15,7 +17,7 @@
  * GNSS-SDR is a Global Navigation Satellite System software-defined receiver.
  * This file is part of GNSS-SDR.
  *
- * Copyright (C) 2010-2020  (see AUTHORS file for a list of contributors)
+ * Copyright (C) 2010-2025  (see AUTHORS file for a list of contributors)
  * SPDX-License-Identifier: GPL-3.0-or-later
  *
  * -----------------------------------------------------------------------------
@@ -24,80 +26,39 @@
 #ifndef GNSS_SDR_GLONASS_L2_CA_DLL_PLL_TRACKING_H
 #define GNSS_SDR_GLONASS_L2_CA_DLL_PLL_TRACKING_H
 
-#include "glonass_ca_dll_pll_tracking_cc.h"
-#include "tracking_interface.h"
-#include <string>
+#include "base_dll_pll_tracking.h"
 
 /** \addtogroup Tracking
+ * Classes for GNSS signal tracking.
  * \{ */
-/** \addtogroup Tracking_adapters
+/** \addtogroup Tracking_adapters tracking_adapters
+ * Wrap GNU Radio blocks for GNSS signal tracking with a TrackingInterface
  * \{ */
 
-
-class ConfigurationInterface;
 
 /*!
  * \brief This class implements a code DLL + carrier PLL tracking loop
+ * block adapter for GLONASS L2 signals
  */
-class GlonassL2CaDllPllTracking : public TrackingInterface
+class GlonassL2CaDllPllTracking : public BaseDllPllTracking
 {
 public:
-    GlonassL2CaDllPllTracking(
-        const ConfigurationInterface* configuration,
+    //! Constructor
+    GlonassL2CaDllPllTracking(const ConfigurationInterface* configuration,
         const std::string& role,
         unsigned int in_streams,
         unsigned int out_streams);
 
-    ~GlonassL2CaDllPllTracking() = default;
-
-    inline std::string role() override
-    {
-        return role_;
-    }
-
-    //! Returns "GLONASS_L1_CA_DLL_PLL_Tracking"
+    //! Returns "GLONASS_L2_CA_DLL_PLL_Tracking"
     inline std::string implementation() override
     {
         return "GLONASS_L2_CA_DLL_PLL_Tracking";
     }
 
-    inline size_t item_size() override
-    {
-        return item_size_;
-    }
-
-    void connect(gr::top_block_sptr top_block) override;
-    void disconnect(gr::top_block_sptr top_block) override;
-    gr::basic_block_sptr get_left_block() override;
-    gr::basic_block_sptr get_right_block() override;
-
-    /*!
-     * \brief Set tracking channel unique ID
-     */
-    void set_channel(unsigned int channel) override;
-
-    /*!
-     * \brief Set acquisition/tracking common Gnss_Synchro object pointer
-     * to efficiently exchange synchronization data between acquisition and tracking blocks
-     */
-    void set_gnss_synchro(Gnss_Synchro* p_gnss_synchro) override;
-
-    void start_tracking() override;
-
-    /*!
-     * \brief Stop running tracking
-     */
-    void stop_tracking() override;
-
 private:
-    glonass_l2_ca_dll_pll_tracking_cc_sptr tracking_sptr_;
-    std::string role_;
-    size_t item_size_;
-    unsigned int channel_;
-    unsigned int in_streams_;
-    unsigned int out_streams_;
+    void configure_tracking_parameters(const ConfigurationInterface* configuration) override;
+    void create_tracking_block() override;
 };
-
 
 /** \} */
 /** \} */